requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
sqlalchemy>=1.4.0
psycopg2-binary>=2.9.1
python-dotenv>=0.19.0
requests>=2.26.0
pytest>=6.2.0
httpx>=0.19.0